# docs/swagger.yaml
...
info:
  contact: {}
paths:
...
  /api/clusters/{clusterName}/proxies/{proxyName}/actions/need-reprov:
    get:
      consumes:
      - application/json
      description: Check if the proxy service for a given cluster and proxy needs
        reprovisioning
      parameters:
      - description: Cluster Name
        in: path
        name: clusterName
        required: true
        type: string
      - description: Proxy Name
        in: path
        name: proxyName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Need reprov!
          schema:
            type: string
        "403":
          description: No valid ACL
          schema:
            type: string
        "500":
          description: No cluster
          schema:
            type: string
        "503":
          description: No reprov needed!" "Not a Valid Server!
          schema:
            type: string
      summary: Check if Proxy Needs Reprovision
      tags:
      - Proxies
  /api/clusters/{clusterName}/proxies/{proxyName}/actions/need-restart:
    get:
      consumes:
      - application/json
      description: Check if the proxy service for a given cluster and proxy needs
        a restart
      parameters:
      - description: Cluster Name
        in: path
        name: clusterName
        required: true
        type: string
      - description: Proxy Name
        in: path
        name: proxyName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Need restart!
          schema:
            type: string
        "403":
          description: No valid ACL
          schema:
            type: string
        "500":
          description: No cluster
          schema:
            type: string
        "503":
          description: No restart needed!" "Not a Valid Server!
          schema:
            type: string
      summary: Check if Proxy Needs Restart
      tags:
      - Proxies
  /api/clusters/{clusterName}/proxies/{proxyName}/actions/provision:
    post:
      consumes:
      - application/json
      description: Provision the proxy service for a given cluster and proxy
      parameters:
      - description: Cluster Name
        in: path
        name: clusterName
        required: true
        type: string
      - description: Proxy Name
        in: path
        name: proxyName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Proxy Service Provisioned
          schema:
            type: string
        "403":
          description: No valid ACL
          schema:
            type: string
        "500":
          description: Cluster Not Found" "Server Not Found
          schema:
            type: string
      summary: Provision Proxy Service
      tags:
      - Proxies
  /api/clusters/{clusterName}/proxies/{proxyName}/actions/start:
    post:
      consumes:
      - application/json
      description: Start the proxy service for a given cluster and proxy
      parameters:
      - description: Cluster Name
        in: path
        name: clusterName
        required: true
        type: string
      - description: Proxy Name
        in: path
        name: proxyName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Proxy Service Started
          schema:
            type: string
        "403":
          description: No valid ACL
          schema:
            type: string
        "500":
          description: Cluster Not Found" "Server Not Found
          schema:
            type: string
      summary: Start Proxy Service
      tags:
      - Proxies
  /api/clusters/{clusterName}/proxies/{proxyName}/actions/stop:
    post:
      consumes:
      - application/json
      description: Stop the proxy service for a given cluster and proxy
      parameters:
      - description: Cluster Name
        in: path
        name: clusterName
        required: true
        type: string
      - description: Proxy Name
        in: path
        name: proxyName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Proxy Service Stopped
          schema:
            type: string
        "403":
          description: No valid ACL
          schema:
            type: string
        "500":
          description: Cluster Not Found" "Server Not Found
          schema:
            type: string
      summary: Stop Proxy Service
      tags:
      - Proxies
  /api/clusters/{clusterName}/proxies/{proxyName}/actions/unprovision:
    post:
      consumes:
      - application/json
      description: Unprovision the proxy service for a given cluster and proxy
      parameters:
      - description: Cluster Name
        in: path
        name: clusterName
        required: true
        type: string
      - description: Proxy Name
        in: path
        name: proxyName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Proxy Service Unprovisioned
          schema:
            type: string
        "403":
          description: No valid ACL
          schema:
            type: string
        "500":
          description: Cluster Not Found" "Server Not Found
          schema:
            type: string
      summary: Unprovision Proxy Service
      tags:
      - Proxies
...
  /api/clusters/{clusterName}/sphinx/indexes:
    get:
      consumes:
      - application/json
      description: Get the Sphinx indexes for a given cluster
      parameters:
      - description: Cluster Name
        in: path
        name: clusterName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Sphinx Indexes
          schema:
            type: string
        "403":
          description: No valid ACL
          schema:
            type: string
        "404":
          description: Something went wrong
          schema:
            type: string
        "500":
          description: Cluster Not Found
          schema:
            type: string
      summary: Get Sphinx Indexes
      tags:
      - Proxies
...
swagger: "2.0"